fix for importing gendiffs.py/embree.py inside houdini
also moved try_decode to lutest_utils.py
